Review: reuse ConstStartsWithVarcharFunction
